+++ payment/src/main/java/com/example/payment/application/dto/ChargeDetailResult.java
@@ -1,6 +1,7 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.ChargeStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -11,8 +12,8 @@ public record ChargeDetailResult(
         UUID chargeId,
         UUID memberId,
         UUID walletId,
-        Long requestedAmount,
-        Long approvedAmount,
+        BigDecimal requestedAmount,
+        BigDecimal approvedAmount,
         String tossBankCode,
         String pgOrderId,
         String pgPaymentKey,
