+++ payment/src/main/java/com/example/payment/application/dto/WithdrawListItemResult.java
@@ -1,14 +1,15 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.WithdrawStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record WithdrawListItemResult(
         UUID withdrawRequestId,
-        Long amount,
-        Long fee,
-        Long actualAmount,
+        BigDecimal amount,
+        BigDecimal fee,
+        BigDecimal actualAmount,
         String maskedBankAccount,
         WithdrawStatus status,
         LocalDateTime requestedAt,
