+++ payment/src/main/java/com/example/payment/presentation/dto/response/ChargeCreateResponse.java
@@ -2,6 +2,7 @@ package com.example.payment.presentation.dto.response;
 
 import com.example.payment.application.dto.ChargeCreateResult;
 import com.example.payment.domain.enumtype.ChargeStatus;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -11,13 +12,12 @@ public record ChargeCreateResponse(
         UUID chargeId,
         UUID walletId,
         String pgOrderId,
-        Long amount,
+        BigDecimal amount,
         ChargeStatus chargeStatus
 ) {
 
     /**
      * application 결과를 presentation 응답 형식으로 변환한다.
-     * 클린아키텍처상 외부 요청형식을 내부실행모델로 바꾸는 것은 외부 계층의 책임이라고 판단한 ai의 판단이 적절하다고 생각함.
      */
     public static ChargeCreateResponse from(ChargeCreateResult result) {
         return new ChargeCreateResponse(
