+++ payment/src/main/java/com/example/payment/presentation/dto/response/PaymentRefundResponse.java
@@ -3,6 +3,7 @@ package com.example.payment.presentation.dto.response;
 import com.example.payment.application.dto.PaymentRefundResult;
 import com.example.payment.domain.enumtype.PaymentRefundStatus;
 import com.example.payment.domain.enumtype.PaymentRefundType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -13,7 +14,7 @@ public record PaymentRefundResponse(
         UUID orderCancelRequestId,
         PaymentRefundStatus refundStatus,
         PaymentRefundType refundType,
-        Long totalRefundAmount,
+        BigDecimal totalRefundAmount,
         List<PaymentRefundItemResponse> itemResults,
         LocalDateTime processedAt
 ) {
