+++ settlement/src/main/java/com/example/settlement/infrastructure/messaging/kafka/contract/SettlementCandidateCreatedMessage.java
@@ -1,5 +1,6 @@
 package com.example.settlement.infrastructure.messaging.kafka.contract;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -8,7 +9,7 @@ public record SettlementCandidateCreatedMessage(
         UUID orderId,
         UUID escrowId,
         UUID sellerMemberId,
-        Long grossAmount,
+        BigDecimal grossAmount,
         Instant releasedAt,
         String confirmationType,
         Instant occurredAt
